+++ ta4j-examples/src/main/java/ta4jexamples/num/CompareNumTypes.java
@@ -1,7 +1,7 @@
 /*******************************************************************************
  *   The MIT License (MIT)
  *
- *   Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2018 Ta4j Organization 
+ *   Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2018 Ta4j Organization
  *   & respective authors (see AUTHORS)
  *
  *   Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -33,32 +33,44 @@ import org.ta4j.core.indicators.helpers.DifferenceIndicator;
 import org.ta4j.core.indicators.helpers.MaxPriceIndicator;
 import org.ta4j.core.indicators.helpers.MinPriceIndicator;
 import org.ta4j.core.num.BigDecimalNum;
-import org.ta4j.core.num.DoubleNum;
 import org.ta4j.core.num.Num;
+import org.ta4j.core.num.PrecisionNum;
+import org.ta4j.core.num.DoubleNum;
 import org.ta4j.core.trading.rules.IsEqualRule;
 import org.ta4j.core.trading.rules.UnderIndicatorRule;
 
 import java.time.ZonedDateTime;
 import java.util.Random;
 
-public class DoubleNumVsBigDecimalNum {
+public class CompareNumTypes {
+
+    private static final int NUMBARS = 10000;
 
     public static void main(String args[]) {
         BaseTimeSeries.SeriesBuilder timeSeriesBuilder = new BaseTimeSeries.SeriesBuilder();
         TimeSeries seriesD = timeSeriesBuilder.withName("Sample Series Double    ").withNumTypeOf(DoubleNum::valueOf).build();
         TimeSeries seriesB = timeSeriesBuilder.withName("Sample Series BigDecimal").withNumTypeOf(BigDecimalNum::valueOf).build();
+        TimeSeries seriesP = timeSeriesBuilder.withName("Sample Series PrecisionNum 32").withNumTypeOf(PrecisionNum::valueOf).build();
+        TimeSeries seriesPH = timeSeriesBuilder.withName("Sample Series PrecisionNum 256").withNumTypeOf(number -> PrecisionNum.valueOf(number.toString(), 256)).build();
 
-        int[] randoms = new Random().ints(100000, 80, 100).toArray();
+        int[] randoms = new Random().ints(NUMBARS, 80, 100).toArray();
         for (int i = 0; i < randoms.length; i++) {
-            ZonedDateTime date = ZonedDateTime.now().minusSeconds(100000 - i);
+            ZonedDateTime date = ZonedDateTime.now().minusSeconds(NUMBARS - i);
             seriesD.addBar(date, randoms[i], randoms[i] + 21, randoms[i] - 21, randoms[i] - 5);
             seriesB.addBar(date, randoms[i], randoms[i] + 21, randoms[i] - 21, randoms[i] - 5);
+            seriesP.addBar(date, randoms[i], randoms[i] + 21, randoms[i] - 21, randoms[i] - 5);
+            seriesPH.addBar(date, randoms[i], randoms[i] + 21, randoms[i] - 21, randoms[i] - 5);
         }
-        test(seriesB);
-        test(seriesD);
+        Num B = PrecisionNum.valueOf(test(seriesB).toString(), 256);
+        Num D = PrecisionNum.valueOf(test(seriesD).toString(), 256);
+        Num P = PrecisionNum.valueOf(test(seriesP).toString(), 256);
+        Num standard = PrecisionNum.valueOf(test(seriesPH).toString(), 256);
+        System.out.println("BigDecimalNum error percent: " + B.minus(standard).dividedBy(standard).multipliedBy(PrecisionNum.valueOf(100)));
+        System.out.println("DoubleNum error percent: " + D.minus(standard).dividedBy(standard).multipliedBy(PrecisionNum.valueOf(100)));
+        System.out.println("PrecisionNum error percent: " + P.minus(standard).dividedBy(standard).multipliedBy(PrecisionNum.valueOf(100)));
     }
 
-    public static void test(TimeSeries series){
+    public static Num test(TimeSeries series) {
         ClosePriceIndicator closePriceIndicator = new ClosePriceIndicator(series);
         RSIIndicator rsi = new RSIIndicator(closePriceIndicator,100);
         MACDIndicator macdIndicator = new MACDIndicator(rsi);
@@ -81,5 +93,6 @@ public class DoubleNumVsBigDecimalNum {
                 "    -Time:   %s ms.\n" +
                 "    -Profit: %s \n" +
                 "    -Bars:   %s\n \n",series.getName(),(end-start),profitResult1, series.getBarCount());
+        return profitResult1;
     }
 }